requirements.txt
fiftyone==0.22.3
matplotlib==3.5.2
natsort==8.2.0
numpy==1.22.3
Pillow==10.1.0
scipy==1.8.1
torch==1.11.0
torchvision==0.12.0